requirements.txt
langchain
langgraph
langchain-core
typing-extensions
python-dotenv
openai
serpapi